Move `get()` and `find()` functions into utils.py
