+++ coc/utils.py
@@ -1,5 +1,7 @@
 import json
 
+from datetime import datetime
+
 
 def to_json(model):
     dct = {}
@@ -31,5 +33,30 @@ def to_json(model):
 
     return json.loads(dct, separators=(',', ':'), ensure_ascii=True)
 
+
+def find(predicate, seq):
+    for element in seq:
+        if predicate(element):
+            return element
+    return None
+
+
+def get(iterable, **attrs):
+    def predicate(elem):
+        for attr, val in attrs.items():
+            nested = attr.split('__')
+            obj = elem
+            for attribute in nested:
+                obj = getattr(obj, attribute)
+
+            if obj != val:
+                return False
+        return True
+
+    return find(predicate, iterable)
+
+
+def from_timestamp(timestamp):
+    return datetime.strptime(timestamp, '%Y%m%dT%H%M%S.000Z')
